+++ packages/contact-center/cc-components/src/index.ts
@@ -3,8 +3,17 @@ import StationLoginComponent from './components/StationLogin/station-login';
 import CallControlComponent from './components/task/CallControl/call-control';
 import IncomingTaskComponent from './components/task/IncomingTask/incoming-task';
 import TaskListComponent from './components/task/TaskList/task-list';
+import OutdialCallComponent from './components/OutdialCall/out-dial-call';
 
-export {UserStateComponent, StationLoginComponent, CallControlComponent, IncomingTaskComponent, TaskListComponent};
+export {
+  UserStateComponent,
+  StationLoginComponent,
+  CallControlComponent,
+  IncomingTaskComponent,
+  TaskListComponent,
+  OutdialCallComponent,
+};
 export * from './components/StationLogin/station-login.types';
 export * from './components/UserState/user-state.types';
 export * from './components/task/task.types';
+export * from './components/OutdialCall/out-dial-call.types';