+++ packages/contact-center/cc-components/src/wc.ts
@@ -4,6 +4,7 @@ import StationLoginComponent from './components/StationLogin/station-login';
 import CallControlComponent from './components/task/CallControl/call-control';
 import IncomingTaskComponent from './components/task/IncomingTask/incoming-task';
 import TaskListComponent from './components/task/TaskList/task-list';
+import OutdialCallComponent from './components/OutdialCall/out-dial-call';
 
 const WebUserState = r2wc(UserStateComponent, {
   props: {
@@ -87,4 +88,9 @@ const WebTaskList = r2wc(TaskListComponent, {
 });
 if (!customElements.get('component-cc-task-list')) {
   customElements.define('component-cc-task-list', WebTaskList);
+}
+
+const WebOutdialCallComponent = r2wc(OutdialCallComponent);
+if (!customElements.get('component-cc-out-dial-call')) {
+  customElements.define('component-cc-out-dial-call', WebOutdialCallComponent);
 }