+++ packages/contact-center/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.28.0-ccwidgets.37](https://github.com/webex/widgets/compare/v1.28.0-ccwidgets.36...v1.28.0-ccwidgets.37) (2025-03-25)
+
+
+### Bug Fixes
+
+* **cc-widgets:** Pause resume record issue fix ([#409](https://github.com/webex/widgets/issues/409)) ([11c9b5e](https://github.com/webex/widgets/commit/11c9b5ecf2745eaf44762afb018c70add4b0014b))
+
+# [1.28.0-ccwidgets.36](https://github.com/webex/widgets/compare/v1.28.0-ccwidgets.35...v1.28.0-ccwidgets.36) (2025-03-25)
+
+
+### Features
+
+* **cc-task:** added agent outdial feature ([#399](https://github.com/webex/widgets/issues/399)) ([cded569](https://github.com/webex/widgets/commit/cded56903e258763df082749702490bd6fc8380c))
+
 # [1.28.0-ccwidgets.35](https://github.com/webex/widgets/compare/v1.28.0-ccwidgets.34...v1.28.0-ccwidgets.35) (2025-03-21)